+++ src/search/search.rs
@@ -7,5 +7,5 @@ pub trait Searcher {
     fn stats(&self) -> &Stats;
     fn best_move(&mut self, pos: &mut BoardState) -> EvaledMove;
     fn best_move_depth(&mut self, pos: &mut BoardState, depth: usize) -> EvaledMove;
-    fn move_time(&mut self, seconds: u64);
+    fn move_time(&mut self, seconds: u128);
 }
